backend/requirements.txt
# CareerRadar Backend Dependencies
# Versions unpinned — pip resolves compatible versions automatically

# FastAPI & Server
fastapi
uvicorn[standard]

# Pydantic & Settings
pydantic
pydantic-settings

# Async HTTP
aiohttp
httpx

# Supabase Database
supabase

# AI Providers
google-genai
anthropic

# Scheduling
apscheduler

# Data Processing
python-dateutil
beautifulsoup4
lxml

# Web Scraping
playwright

# Form/File uploads
python-multipart

# CV Processing
PyPDF2

# Utilities
python-dotenv

# Error Tracking
sentry-sdk[fastapi]

# ML & Vector Operations (CareerRadar Upgrade)
sentence-transformers
numpy
scikit-learn

# System Monitoring
psutil
